organized files according to PR#1
added isostasy_pratt to functions. Added exampled to the example folder
