+++ harmonica/__init__.py
@@ -5,6 +5,7 @@ from . import datasets
 from . import synthetic
 from .io import load_icgem_gdf
 from .isostasy import isostasy_airy
+from .isostasy import isostasy_pratt
 from .gravity_corrections import bouguer_correction
 from .forward.point_mass import point_mass_gravity
 from .forward.tesseroid import tesseroid_gravity
